backend/requirements.txt
# Anupalan Mitra — Backend Dependencies
# Offline embeddings via sentence-transformers (no API rate limits)

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
google-generativeai>=0.8.3
chromadb>=0.5.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
torch>=2.0.0
